Paginación

+++ resources/views/users/index.blade.php
@@ -26,6 +26,7 @@
 
             @if (count($errors) > 0 )
               <div class="alert alert-danger">
+                <h2>¡Ha ocurrido un error!</h2>
                 <ul>
                     @foreach ($errors->all() as $error)
                       <li>{{ $error }}</li>
@@ -85,7 +86,7 @@
                       <h3 class="h4">Listado de Usuarios</h3>
                     </div>
                     <div class="card-body">
-                      <table class="table">
+                      <table class="table" id="datatable">
                         <thead>
                           <tr>
                             <th>#</th>
@@ -105,7 +106,7 @@
                                   <a href="/usuario/{{ $user->id }}" class="btn btn-sm btn-primary" title="Editar">
                                     <i class="fa fa-pencil"></i>
                                   </a>
-                                  <a href="/usuario/{{ $user->id }}/eliminar" class="btn btn-sm btn-danger" title="Eliminar">
+                                  <a href="/usuario/{{ $user->id }}/eliminar" class="btn btn-sm btn-danger" onclick="return confirm('¿Está seguro de eliminar este registro?')" title="Eliminar">
                                     <i class="fa fa-trash"></i>
                                   </a>                                
                               </td>
@@ -121,4 +122,8 @@
             </div>
           </section>   
         </div>                 
+@endsection
+@section('footer')
+<script src='https://cdn.datatables.net/1.10.7/js/jquery.dataTables.min.js'></script>
+<script  src="/js/datatable_custom.js"></script>
 @endsection
